requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
dnspython==2.8.0
fastapi==0.124.0
filelock==3.20.0
h11==0.16.0
idna==3.11
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
motor==3.7.1
msgpack==1.1.2
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
platformdirs==4.5.1
py-serializable==2.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
pyparsing==3.2.5
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
tomli==2.3.0
tomli_w==1.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
